+++ CHANGELOG.rst
@@ -1,6 +1,8 @@
 1.2.3
 -----
 
+* Added :class:`.Percent` computation. (#490)
+* Added Ben Welsh to AUTHORS.
 * :meth:`.Table.__init__` no longer fails on duplicate columns. Instead it renames them and throws warning. (#484)
 * :meth:`.Table.merge` now takes a ``column_names`` argument to specify columns included in new table. (#481)
 * :meth:`.Table.select` now accepts a single column name as a key.